+++ packages/scrollable-anchors/client/example/index.js
@@ -1,8 +1,6 @@
 import React, { useState } from 'react'
 import { observer } from 'startupjs'
-import { useHistory } from '@startupjs/app'
-import { Span, Br, Div, H1, H3, H4, Divider, Row, Button } from '@startupjs/ui'
-import { faArrowLeft } from '@fortawesome/free-solid-svg-icons'
+import { Span, Br, Div, H1, H3, H4, Divider, Row, Button, BackButton } from '@startupjs/ui'
 import { ScrollableArea, Anchor } from '../components'
 import { scrollTo } from '../helpers'
 import './index.styl'
@@ -18,7 +16,6 @@ const LOREM = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do e
 
 function Example () {
   const [extraContentVisibility, setExtraContentVisibility] = useState(false)
-  const history = useHistory()
 
   function scrollToAnchor ({ anchorId, areaId, offset }) {
     scrollTo({ anchorId, areaId, offset })
@@ -26,12 +23,7 @@ function Example () {
 
   return pug`
     Div.wrapper
-      Button.backButton(
-        icon=faArrowLeft
-        size='m'
-        variant='text'
-        onPress=() => history.goBack()
-      ) 
+      BackButton
     Div.anchors
       Button.extraButton(
         onPress=()=> setExtraContentVisibility(!extraContentVisibility)
