+++ packages/scrollable-anchors/client/example/index.styl
@@ -47,12 +47,3 @@
 
 .wrapper
   background-color $UI.colors.white
-
-.backButton
-  max-width 2u
-  margin-top 4u
-  margin-left 2u
-
-  +tablet()
-    padding-left 3u
-    padding-right 1u
