+++ packages/ui/components/BackButton/index.styl
@@ -1,0 +1,8 @@
+.backButton
+  max-width 2u
+  margin-top 4u
+  margin-left 2u
+
+  +tablet()
+    padding-left 3u
+    padding-right 1u
