+++ styleguide/auth/Layout.js
@@ -1,22 +1,13 @@
 import React from 'react'
 import { ScrollView, Image } from 'react-native'
 import { observer, u } from 'startupjs'
-import { useHistory } from '@startupjs/app'
-import { Div, Button } from '@startupjs/ui'
-import { faArrowLeft } from '@fortawesome/free-solid-svg-icons'
+import { Div, BackButton } from '@startupjs/ui'
 import './index.styl'
 
 export default observer(function Layout ({ children }) {
-  const history = useHistory()
-
   return pug`
     ScrollView.root
-      Button.backButton(
-        icon=faArrowLeft
-        size='m'
-        variant='text'
-        onPress=() => history.goBack()
-      ) 
+      BackButton.backButton
       Div.logo
         Image(
           resizeMode='contain'
