+++ styleguide/auth/index.styl
@@ -41,10 +41,6 @@
   margin-top 2u
 
 .backButton
-  max-width 2u
-  margin-top 4u
-  margin-left 2u
-
   +tablet()
     padding-left 3u
     padding-right 1u
