+++ styleguide/main/pages/PProfile/index.styl
@@ -12,12 +12,3 @@
 
 .main
   background-color $UI.colors.white
-
-.backButton
-  max-width 2u
-  margin-top 3u
-  margin-left 3u
-
-  +tablet()
-    padding-left 3u
-    padding-right 1u
